Allow hostonlynets networks of any name to be created
With VirtualBox v7 on macOS we can create networks of 'hostonlynets' with a
new CLI experience that allows any name for the created network.

We leverage this new possibility, while still retaining the former feature
with networks of 'hostonlyif' type, when the name of a non-existing and
to-be-created network could possibly be guessed in advance, and if correctly
guessed, then everything was fine.